Safari-611.1.10/LayoutTests/http/tests/websocket/tests/hybi/websocket-allowed-setting-cookie-as-third-party-expected.txt

Tests WebSocket Set-Cookie behavior for third-parties with existing cookies.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.location.host is "127.0.0.1:8000"

Setting third-party cookie 'foo' through cross-origin WebSocket handshake.

Opening localhost third-party iframe to check its cookies.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Cookies are: setAsFirstParty = value foo = bar
